Added size editor and spawnicon
The size editor does work but it doesn't let you spawn a Safe Space
using your settings yet. I also had to mess with the codebase a bit to
get it to work in the preview window so I need to clean-up the code that
gets the dimensions/lighting/etc. The code for the sliders also isn't
very clean so that will need clean-up too before release.

- Moved dimension/lighting Get methods to SafeSpace library
- Added option menu with button to open editor
- Added editor with sliders to show a user custom-sized Safe Space in
real time
- Added offset to interior cubes so the local origin is in the middle
rather than top right corner
- Added spawnicon